fix(tvix/eval): ensure all evaluated thunks are correctly memoized
This fixes a very complicated bug (b/246), and likely introduces a
handful of less complicated new bugs in the process.

What was the problem?
=====================

Previously, when evaluating a thunk, we had a code path that looked
like this:

    match *thunk {
        ThunkRepr::Evaluated(Value::Thunk(ref inner_thunk)) => {
            let inner_repr = inner_thunk.0.borrow().clone();
            drop(thunk);
            self.0.replace(inner_repr);
        }
        /* ... */
    }

This code path created a copy of the inner `ThunkRepr` of a nested
thunk, and moved that copy into the `ThunkRepr` of the parent.

The effect of this was that the original `ThunkRepr` (unforced!) lived
on in the original thunk, without the memoization of the subsequent
forcing applying to it.

This had the result that Tvix would repeatedly evaluate these thunks
without ever memoizing them, if they occured repeatedly as shared
inner thunks. Most notably, this would *always* occur when
builtins.import was used.

What's the solution?
====================

I have completely rewritten `Thunk::force_trampoline_self` to make all
flows that can occur in it explicit. I have also removed the outer
loop inside of that function, and resorted to more use of trampolining
instead.

The function is now well-commented and it should be possible to read
it from top-to-bottom and get a general sense of what is going on,
though the trampolining itself (which is implemented in the VM) needs
to be at least partially understood for this.

What's the new problem(s)?
==========================

One new (known) problem is that we have to construct `Error` instances
in all error types here, but we do not have spans available in some
thunk-related situations. Due to b/238 we cannot ask the VM for an
arbitrary span from the callsite leading to the force. This means that
there are now code paths where, under certain conditions, causing an
evaluation error during thunk forcing will panic.

To fix this we will need to investigate and fix b/238, and/or add a
span tracking mechanism to thunks themselves.

What other impacts does this have?
==================================

As of this commit, this introduces some kind of new thunking issue
that becomes apparent when evaluating more involved expressions in
nixpkgs (e.g. `stdenv`), where some thunks remain unforced. This could
be another bug surfacing as a result of this commit, and is not
something I can tackle immediately in this commit.

As usual, we will try to isolate a test case for that behaviour and
then fix it.

Additionally, this makes eval around 30% faster than before when
evaluating very small chunks of nixpkgs. It probably has an
exponentially larger impact, but because of the above we can't confirm
that (yet).

Change-Id: I738525bad8bc5ede5d8c737f023b14b8f4160612