fix: make depot-gcroot derivation a plain list of out paths
This just creates a text file (similar to the propagated deps file)
which lists the outpaths of all targets.

It's unclear why the previous fix didn't work, but it was done solely
based on docs not on reading code. In general we don't really need a
setup hook though and nobody remembers why that was the solution we
used anyways, so lets remove it.

Change-Id: I2a945925f4fa56b272aa44c29229b6a1aefa80e6