chore(3p/sources): bump home-manager
Currently there is an issue in nmd which prevents it from working with
Nix 2.3 (it passes --store dummy://) which is why we haven't updated
home-manager for a while. Since home-manager pins nmd in a
non-overrideable way, we need to prevent its use until an upstream patch
has landed and propagated to home-manager (fix seems trivial, I'll try
to submit it soon).

This can be done by simply disabling manual.* — in our case only
manual.manpages was used which is enabled by default.

Change-Id: Ia1447549c0f97aa754ac1842eb453e95838c00c5