refactor(nix-compat): Use `StorePath` in `Derivation`
This makes derivations more type safe, and conforms with how things are
done in recent versions of Nix (since 2.4).

Note that we also need `PreDerivation` for the hash modulo computation.
This adds some complexity but only internally (we hope); code that
doesn't care about the details of that downstream should not notice.

Change-Id: Iaa6bdbb61a0cfc4cd8c3f8c82e3f8a669130f79e