refactor(tvix/store): remove ChunkService
Chunking things is an implementation detail of each Blobstore. It
currently is not visible in the gRPC interface either. It shouldn't
bleed into everything.

Let the BlobService trait provide `open_read` and `open_write` methods,
which return handles providing io::Read or io::Write, and leave the
details up to the implementation.

This means, our custom BlobReader module can go away, and all the
chunking bits in there, too.

In the future, we might still want to add more chunking-aware syncing,
but as a syncing strategy some stores can expose, not as a fundamental
rotocol component.

This currently needs a custom fork of tokio, providing "SyncReadStream",
as we need to return a sync reader for an AsyncRead. Another option
would be to make the blobstore trait itself return an AsyncRead, but
then every consumer would need to know what to do with the async-ness.

Change-Id: I5969ebbc4c0e1ceece47981be3b9e7cfb3f59ad0