feat(3p/ast-grep): generate config for tree-sitter grammars
Adds a helper function at `ast-grep.configWithLanguages` that can
generate an ast-grep configuration file with the provided tree-sitter
grammars from nixpkgs. See the doc comment for detailed usage
instructions.

ast-grep currently does not provide an option to override or merge
config files, so for most commands this only works if the resulting
configuration is copied to `sgconfig.yml` in the project root.

Future work might include writing a patch for ast-grep that improves
this language configuration handling.

Change-Id: Iae2de579f4e7fc019a7d60fbcbfd773edacb44c3